Classe SoudnControler criada para armazenar e configurar sons do jogo
